requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2
python-dotenv
requests
httpx
passlib
aiohttp
starlette
